requirements.txt
flask
flask-cors
mysql-connector-python
sqlalchemy
click
pydantic
email-validator
bcrypt
pyjwt
ua-parser

